backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
groq==0.4.2
chromadb==0.4.24
sentence-transformers==2.5.1
pymupdf==1.23.26
python-dotenv==1.0.1
jinja2==3.1.3